moved Plot into own class, renamed it to GardenPlot, and put it in new package. also moved GardenPlotApi and GardenPlotBorders in the same package
