+++ src/main/java/at/hannibal2/skyhanni/features/garden/plot/GardenPlotBorders.kt
@@ -1,9 +1,10 @@
-package at.hannibal2.skyhanni.features.garden
+package at.hannibal2.skyhanni.features.garden.plot
 
 import at.hannibal2.skyhanni.api.event.HandleEvent
 import at.hannibal2.skyhanni.events.minecraft.KeyDownEvent
 import at.hannibal2.skyhanni.events.minecraft.SkyHanniRenderWorldEvent
-import at.hannibal2.skyhanni.features.garden.GardenPlotApi.renderPlot
+import at.hannibal2.skyhanni.features.garden.GardenApi
+import at.hannibal2.skyhanni.features.garden.plot.GardenPlotApi.renderPlot
 import at.hannibal2.skyhanni.skyhannimodule.SkyHanniModule
 import at.hannibal2.skyhanni.utils.LocationUtils.distanceSqToPlayer
 import at.hannibal2.skyhanni.utils.LorenzColor
@@ -37,7 +38,7 @@ object GardenPlotBorders {
         )
     }
 
-    private fun getClosestPlot(): GardenPlotApi.Plot? =
+    private fun getClosestPlot(): GardenPlot? =
         GardenPlotApi.plots.minByOrNull { it.middle.distanceSqToPlayer() }
 
     private fun isEnabled() = GardenApi.inGarden() && config.plotBorderKey != GLFW.GLFW_KEY_UNKNOWN
